+++ lib/galaxy/model/dataset_collections/registry.py
