+++ lib/galaxy/model/dataset_collections/structure.py
