+++ lib/galaxy/model/dataset_collections/subcollections.py
