+++ lib/galaxy/model/dataset_collections/type_description.py
