+++ lib/galaxy/model/dataset_collections/types/list.py
