+++ lib/galaxy/model/dataset_collections/types/paired.py
