+++ lib/galaxy/tools/actions/upload.py
@@ -2,8 +2,8 @@ import json
 import logging
 import os
 
-from galaxy.dataset_collections.structure import UninitializedTree
 from galaxy.exceptions import RequestParameterMissingException
+from galaxy.model.dataset_collections.structure import UninitializedTree
 from galaxy.tools.actions import upload_common
 from galaxy.util import ExecutionTimer
 from galaxy.util.bunch import Bunch
